+++ app/_/components/search/type.ts
@@ -2,3 +2,6 @@ export type SuggestionsListProps = {
     suggestions: string[]
     onSuggestionClick: (suggestion: string) => void
 }
+export type SearchBarProps = {
+    onSearch: (keyword: string) => void
+}
